update HockeyApp IOS SDK to 3.8.5 and remove bitcode warning as this version of the sdk supports bitcode
