+++ src/ios/HockeySDK.embeddedframework/HockeySDK.framework/Modules/module.modulemap
@@ -1,0 +1,18 @@
+framework module HockeySDK {
+  umbrella header "HockeySDK.h"
+  
+  export *
+  module * { export * }
+  
+  link framework "AssetsLibrary"
+  link framework "CoreText"
+  link framework "CoreGraphics"
+  link framework "Foundation"
+  link framework "MobileCoreServices"
+  link framework "QuartzCore"
+  link framework "QuickLook"
+  link framework "Security"
+  link framework "SystemConfiguration"
+  link framework "UIKit"
+  link "c++"
+}
